requirements.txt
flask
pyyaml
aiohttp
requests
